+++ src/background.js
@@ -1,6 +1,14 @@
-import {iconListeners} from "./background/icon/iconMain";
-import {mainActionListener} from "./background/action/actionMain";
+import {iconListeners} from "./background/icon/icon";
+import {clickActionListener} from "./background/export/exportOnce";
+import {exportAllThreadsListener} from "./background/export/exportAll";
 
-mainActionListener();
+clickActionListener();
 iconListeners();
+exportAllThreadsListener(); // Needs to be after iconListeners and clickActionListener to work properly
 
+
+chrome.runtime.onInstalled.addListener(function () {
+  chrome.storage.sync.set({displayModalUpdate: true}, function () {
+    console.log("Last update modal will be displayed");
+  });
+});