+++ src/background/export/exportOnce.js
@@ -1,7 +1,7 @@
 /***
  * This is the code for the extension to run when the icon is clicked
  */
-export function mainActionListener() {
+export function clickActionListener() {
   chrome.action.onClicked.addListener(async (tab) => {
     chrome.scripting.executeScript({
       target: {tabId: tab.id},